refactor: move OfferCode service types to Domain package
Aligns with the established pattern where service protocols (Authentication,
Subscription, Notification, etc.) live in TownCrierDomain as ports, with
concrete implementations in TownCrierData as adapters. This lets the
Presentation layer depend on OfferCodeService without pulling in TownCrierData.
