+++ issues.jsonl
@@ -90,7 +90,7 @@
 {"id":"tc-l43","title":"Handle 429 with skip-then-stop","description":"Spec: docs/superpowers/specs/2026-04-05-polling-resilience-design.md#3-handle-429-with-skip-then-stop","status":"closed","priority":1,"issue_type":"task","assignee":"AmeDe","owner":"amy@salter.uk","created_at":"2026-04-05T11:31:23Z","created_by":"AmeDe","updated_at":"2026-04-06T07:33:29Z","closed_at":"2026-04-05T13:05:05Z","close_reason":"Closed","dependencies":[{"issue_id":"tc-l43","depends_on_id":"tc-69h","type":"blocks","created_at":"2026-04-05T12:31:31Z","created_by":"AmeDe","metadata":"{}"}],"dependency_count":1,"dependent_count":0,"comment_count":0}
 {"id":"tc-9rcg.17","title":"Web RedeemOfferCode component + Settings integration","description":"Spec: docs/specs/offer-codes.md · Plan: docs/superpowers/plans/2026-04-18-offer-codes.md (Task I3)","status":"open","priority":2,"issue_type":"task","owner":"amy@salter.uk","created_at":"2026-04-18T18:18:33Z","created_by":"AmyDe","updated_at":"2026-04-18T18:18:33Z","labels":["api","cli","infra","ios","web"],"dependencies":[{"issue_id":"tc-9rcg.17","depends_on_id":"tc-9rcg","type":"parent-child","created_at":"2026-04-18T19:18:33Z","created_by":"AmyDe","metadata":"{}"},{"issue_id":"tc-9rcg.17","depends_on_id":"tc-9rcg.16","type":"blocks","created_at":"2026-04-18T19:19:35Z","created_by":"AmyDe","metadata":"{}"}],"dependency_count":1,"dependent_count":1,"comment_count":0}
 {"id":"tc-9rcg.16","title":"Web useRedeemOfferCode hook + spy + tests","description":"Spec: docs/specs/offer-codes.md · Plan: docs/superpowers/plans/2026-04-18-offer-codes.md (Task I2)","status":"open","priority":2,"issue_type":"task","owner":"amy@salter.uk","created_at":"2026-04-18T18:18:32Z","created_by":"AmyDe","updated_at":"2026-04-18T18:18:32Z","labels":["api","cli","infra","ios","web"],"dependencies":[{"issue_id":"tc-9rcg.16","depends_on_id":"tc-9rcg","type":"parent-child","created_at":"2026-04-18T19:18:32Z","created_by":"AmyDe","metadata":"{}"},{"issue_id":"tc-9rcg.16","depends_on_id":"tc-9rcg.15","type":"blocks","created_at":"2026-04-18T19:19:34Z","created_by":"AmyDe","metadata":"{}"}],"dependency_count":1,"dependent_count":1,"comment_count":0}
-{"id":"tc-9rcg.15","title":"Web redeem offer code API client + types","description":"Spec: docs/specs/offer-codes.md · Plan: docs/superpowers/plans/2026-04-18-offer-codes.md (Task I1)","status":"in_progress","priority":2,"issue_type":"task","owner":"amy@salter.uk","created_at":"2026-04-18T18:18:31Z","created_by":"AmyDe","updated_at":"2026-04-18T20:04:40Z","started_at":"2026-04-18T20:04:40Z","labels":["api","cli","infra","ios","web"],"dependencies":[{"issue_id":"tc-9rcg.15","depends_on_id":"tc-9rcg","type":"parent-child","created_at":"2026-04-18T19:18:30Z","created_by":"AmyDe","metadata":"{}"},{"issue_id":"tc-9rcg.15","depends_on_id":"tc-9rcg.9","type":"blocks","created_at":"2026-04-18T19:19:34Z","created_by":"AmyDe","metadata":"{}"}],"dependency_count":1,"dependent_count":1,"comment_count":0}
+{"id":"tc-9rcg.15","title":"Web redeem offer code API client + types","description":"Spec: docs/specs/offer-codes.md · Plan: docs/superpowers/plans/2026-04-18-offer-codes.md (Task I1)","status":"closed","priority":2,"issue_type":"task","owner":"amy@salter.uk","created_at":"2026-04-18T18:18:31Z","created_by":"AmyDe","updated_at":"2026-04-18T20:17:29Z","started_at":"2026-04-18T20:04:40Z","closed_at":"2026-04-18T20:17:29Z","close_reason":"Closed","labels":["api","cli","infra","ios","web"],"dependencies":[{"issue_id":"tc-9rcg.15","depends_on_id":"tc-9rcg","type":"parent-child","created_at":"2026-04-18T19:18:30Z","created_by":"AmyDe","metadata":"{}"},{"issue_id":"tc-9rcg.15","depends_on_id":"tc-9rcg.9","type":"blocks","created_at":"2026-04-18T19:19:34Z","created_by":"AmyDe","metadata":"{}"}],"dependency_count":1,"dependent_count":1,"comment_count":0}
 {"id":"tc-9rcg.14","title":"iOS RedeemOfferCodeView + Settings integration","description":"Spec: docs/specs/offer-codes.md · Plan: docs/superpowers/plans/2026-04-18-offer-codes.md (Task H3)","status":"open","priority":2,"issue_type":"task","owner":"amy@salter.uk","created_at":"2026-04-18T18:18:29Z","created_by":"AmyDe","updated_at":"2026-04-18T18:18:29Z","labels":["api","cli","infra","ios","web"],"dependencies":[{"issue_id":"tc-9rcg.14","depends_on_id":"tc-9rcg","type":"parent-child","created_at":"2026-04-18T19:18:28Z","created_by":"AmyDe","metadata":"{}"},{"issue_id":"tc-9rcg.14","depends_on_id":"tc-9rcg.13","type":"blocks","created_at":"2026-04-18T19:19:33Z","created_by":"AmyDe","metadata":"{}"}],"dependency_count":1,"dependent_count":1,"comment_count":0}
 {"id":"tc-9rcg.13","title":"iOS RedeemOfferCodeViewModel + fake + tests","description":"Spec: docs/specs/offer-codes.md · Plan: docs/superpowers/plans/2026-04-18-offer-codes.md (Task H2)","status":"in_progress","priority":2,"issue_type":"task","owner":"amy@salter.uk","created_at":"2026-04-18T18:18:28Z","created_by":"AmyDe","updated_at":"2026-04-18T20:15:27Z","started_at":"2026-04-18T20:15:27Z","labels":["api","cli","infra","ios","web"],"dependencies":[{"issue_id":"tc-9rcg.13","depends_on_id":"tc-9rcg","type":"parent-child","created_at":"2026-04-18T19:18:27Z","created_by":"AmyDe","metadata":"{}"},{"issue_id":"tc-9rcg.13","depends_on_id":"tc-9rcg.12","type":"blocks","created_at":"2026-04-18T19:19:32Z","created_by":"AmyDe","metadata":"{}"}],"dependency_count":1,"dependent_count":1,"comment_count":0}
 {"id":"tc-9rcg.12","title":"iOS OfferCodeService protocol + HTTP impl","description":"Spec: docs/specs/offer-codes.md · Plan: docs/superpowers/plans/2026-04-18-offer-codes.md (Task H1)","status":"closed","priority":2,"issue_type":"task","owner":"amy@salter.uk","created_at":"2026-04-18T18:18:26Z","created_by":"AmyDe","updated_at":"2026-04-18T20:14:22Z","started_at":"2026-04-18T19:59:45Z","closed_at":"2026-04-18T20:14:22Z","close_reason":"Closed","labels":["api","cli","infra","ios","web"],"dependencies":[{"issue_id":"tc-9rcg.12","depends_on_id":"tc-9rcg","type":"parent-child","created_at":"2026-04-18T19:18:26Z","created_by":"AmyDe","metadata":"{}"},{"issue_id":"tc-9rcg.12","depends_on_id":"tc-9rcg.9","type":"blocks","created_at":"2026-04-18T19:19:32Z","created_by":"AmyDe","metadata":"{}"}],"dependency_count":1,"dependent_count":1,"comment_count":0}
