+++ mobile/ios/town-crier-tests/Sources/Features/OfferCodeErrorTests.swift
@@ -1,5 +1,6 @@
 import Foundation
 import Testing
+import TownCrierDomain
 
 @testable import TownCrierData
